requirements.txt
google-genai
pydantic
python-dotenv
pytest
pytest-mock
Pillow
Pillow>=10.0.0
requests
websocket-client
rembg[cpu]
scikit-learn
numpy
